Remove test duplication
This merges the ObjectValue and MutableObjectValue tests. They used to be different since we had ObjectValue and ObjectValueBuilder. The PR also makes an effort to remove test duplication, but makes no guarantees here.
